Replace react-query in breadcrumbs with new useBreadcrumbs() hook
