+++ src/features/breadcrumbs/components/BreadcrumbTrail.tsx
@@ -3,24 +3,13 @@ import makeStyles from '@mui/styles/makeStyles';
 import NavigateNextIcon from '@mui/icons-material/NavigateNext';
 import NextLink from 'next/link';
 import { Theme } from '@mui/material/styles';
-import { useQuery } from 'react-query';
 import { Breadcrumbs, Link, Typography, useMediaQuery } from '@mui/material';
-import { NextRouter, useRouter } from 'next/router';
 
 import { Breadcrumb } from 'utils/types';
-import getBreadcrumbs from '../../../utils/fetching/getBreadcrumbs';
 import { Msg } from 'core/i18n';
 
 import messageIds from '../l10n/messageIds';
-
-const getQueryString = function (router: NextRouter): string {
-  // Only use parameters that are part of the path (e.g. [personId])
-  // and not ones that are part of the actual querystring (e.g. ?filter_*)
-  return Object.entries(router.query)
-    .filter(([key]) => router.pathname.includes(`[${key}]`))
-    .map(([key, val]) => `${key}=${val}`)
-    .join('&');
-};
+import useBreadcrumbElements from '../hooks/useBreadcrumbs';
 
 const useStyles = makeStyles<Theme, { highlight?: boolean }>((theme) =>
   createStyles({
@@ -62,20 +51,10 @@ const BreadcrumbTrail = ({
   highlight?: boolean;
 }): JSX.Element | null => {
   const classes = useStyles({ highlight });
-  const router = useRouter();
-  const path = router.pathname;
-  const query = getQueryString(router);
-  const breadcrumbsQuery = useQuery(
-    ['breadcrumbs', path, query],
-    getBreadcrumbs(path, query)
-  );
+  const breadcrumbs = useBreadcrumbElements();
   const smallScreen = useMediaQuery('(max-width:700px)');
   const mediumScreen = useMediaQuery('(max-width:960px)');
   const largeScreen = useMediaQuery('(max-width:1200px)');
-
-  if (!breadcrumbsQuery.isSuccess) {
-    return <div />;
-  }
 
   const getLabel = (crumb: Breadcrumb) => {
     if (crumb.labelMsg) {
@@ -97,8 +76,8 @@ const BreadcrumbTrail = ({
         maxItems={smallScreen ? 2 : mediumScreen ? 4 : largeScreen ? 6 : 10}
         separator={<NavigateNextIcon fontSize="small" />}
       >
-        {breadcrumbsQuery.data.map((crumb, index) => {
-          if (index < breadcrumbsQuery.data.length - 1) {
+        {breadcrumbs.map((crumb, index) => {
+          if (index < breadcrumbs.length - 1) {
             return (
               <NextLink key={crumb.href} href={crumb.href} passHref>
                 <Link
